+++ packages/keyring-snap-bridge/src/SnapKeyring.ts
@@ -38,6 +38,11 @@ import {
   type InternalAccount,
 } from '@metamask/keyring-internal-api';
 import { KeyringInternalSnapClient } from '@metamask/keyring-internal-snap-client';
+import {
+  type GetSelectedAccountsResponse,
+  GetSelectedAccountsRequestStruct,
+  SnapManageAccountsMethod,
+} from '@metamask/keyring-snap-sdk';
 import type { AccountId, JsonRpcRequest } from '@metamask/keyring-utils';
 import { strictMask } from '@metamask/keyring-utils';
 import type { ExtractEventPayload } from '@metamask/messenger';
@@ -184,6 +189,11 @@ export class SnapKeyring {
   }>;
 
   /**
+   * Mapping between Snap IDs and the selected accounts.
+   */
+  readonly #selectedAccounts: Map<SnapId, AccountId[]>;
+
+  /**
    * Mapping between request IDs and their deferred promises.
    */
   readonly #requests: SnapIdMap<{
@@ -238,6 +248,7 @@ export class SnapKeyring {
     this.#options = new SnapIdMap();
     this.#callbacks = callbacks;
     this.#isAnyAccountTypeAllowed = isAnyAccountTypeAllowed;
+    this.#selectedAccounts = new Map();
   }
 
   /**
@@ -498,6 +509,21 @@ export class SnapKeyring {
   }
 
   /**
+   * Handle a Get Selected Accounts method call from a Snap.
+   *
+   * @param snapId - Snap ID.
+   * @param message - Method call message.
+   * @returns The selected accounts.
+   */
+  async #handleGetSelectedAccounts(
+    snapId: SnapId,
+    message: SnapMessage,
+  ): Promise<GetSelectedAccountsResponse> {
+    assert(message, GetSelectedAccountsRequestStruct);
+    return this.#selectedAccounts.get(snapId) ?? [];
+  }
+
+  /**
    * Handle an Request Approved event from a Snap.
    *
    * @param snapId - Snap ID.
@@ -699,6 +725,10 @@ export class SnapKeyring {
 
       case `${KeyringEvent.AccountTransactionsUpdated}`: {
         return this.#handleAccountTransactionsUpdated(snapId, message);
+      }
+
+      case `${SnapManageAccountsMethod.GetSelectedAccounts}`: {
+        return this.#handleGetSelectedAccounts(snapId, message);
       }
 
       default:
@@ -1498,6 +1528,48 @@ export class SnapKeyring {
   }
 
   /**
+   * Update the in-memory selected accounts map.
+   *
+   * @param accounts - The accounts to update the map with.
+   */
+  #updateSelectedAccountsMap(accounts: AccountId[]): void {
+    const selectedAccounts = this.#selectedAccounts;
+    selectedAccounts.clear();
+    for (const account of accounts) {
+      const snapId = this.#accounts.getSnapId(account);
+      if (!snapId) {
+        continue;
+      }
+      const snapAccounts = selectedAccounts.get(snapId) ?? [];
+      snapAccounts.push(account);
+      selectedAccounts.set(snapId, snapAccounts);
+    }
+  }
+
+  /**
+   * Set the selected accounts.
+   *
+   * @param accounts - The accounts to set as selected.
+   */
+  async setSelectedAccounts(accounts: AccountId[]): Promise<void> {
+    this.#updateSelectedAccountsMap(accounts);
+    const entries = [...this.#selectedAccounts.entries()];
+    await Promise.all(
+      entries.map(async ([snapId, accountIds]) => {
+        try {
+          await this.#snapClient
+            .withSnapId(snapId)
+            .setSelectedAccounts(accountIds);
+        } catch (error: any) {
+          console.error(
+            `Failed to set selected accounts for ${snapId} snap: '${error.message}'`,
+          );
+        }
+      }),
+    );
+  }
+
+  /**
    * Get the Snap associated with the given Snap ID.
    *
    * @param snapId - Snap ID.
@@ -1535,6 +1607,35 @@ export class SnapKeyring {
     );
   }
 
+  #transformToInternalAccount(
+    account: KeyringAccount,
+    snapId: SnapId,
+  ): InternalAccount {
+    const snap = this.#getSnapMetadata(snapId);
+
+    return {
+      ...account,
+      // TODO: Do not convert the address to lowercase.
+      //
+      // This is a workaround to support the current UI which expects the
+      // account address to be lowercase. This workaround should be removed
+      // once we migrated the UI to use the account ID instead of the account
+      // address.
+      //
+      // NOTE: We convert the address only for EVM accounts, see
+      // `normalizeAccountAddress`.
+      address: normalizeAccountAddress(account),
+      metadata: {
+        name: '',
+        importTime: 0,
+        keyring: {
+          type: this.type,
+        },
+        ...(snap !== undefined && { snap }),
+      },
+    };
+  }
+
   /**
    * Return an internal account object for a given address.
    *
@@ -1542,41 +1643,26 @@ export class SnapKeyring {
    * @returns An internal account object for the given address.
    */
   getAccountByAddress(address: string): InternalAccount | undefined {
-    const accounts = this.listAccounts();
-    return accounts.find(({ address: accountAddress }) =>
+    const accounts = [...this.#accounts.values()];
+    const account = accounts.find(({ account: { address: accountAddress } }) =>
       equalsIgnoreCase(accountAddress, address),
     );
+
+    return account
+      ? this.#transformToInternalAccount(account.account, account.snapId)
+      : undefined;
   }
 
   /**
    * List all Snap keyring accounts.
+   * This method is expensive on mobile devices and could takes tens or hundreds of milliseconds to complete.
+   * Use with caution.
    *
    * @returns An array containing all Snap keyring accounts.
    */
   listAccounts(): InternalAccount[] {
-    return [...this.#accounts.values()].map(({ account, snapId }) => {
-      const snap = this.#getSnapMetadata(snapId);
-      return {
-        ...account,
-        // TODO: Do not convert the address to lowercase.
-        //
-        // This is a workaround to support the current UI which expects the
-        // account address to be lowercase. This workaround should be removed
-        // once we migrated the UI to use the account ID instead of the account
-        // address.
-        //
-        // NOTE: We convert the address only for EVM accounts, see
-        // `normalizeAccountAddress`.
-        address: normalizeAccountAddress(account),
-        metadata: {
-          name: '',
-          importTime: 0,
-          keyring: {
-            type: this.type,
-          },
-          ...(snap !== undefined && { snap }),
-        },
-      };
-    });
+    return [...this.#accounts.values()].map(({ account, snapId }) =>
+      this.#transformToInternalAccount(account, snapId),
+    );
   }
 }